requirements.txt
numpy
pandas
matplotlib
seaborn
mplfinance
Pillow
scikit-learn
tensorflow
opencv-python
scikit-image
torch
torchvision
timm
